Moderation action logging
When users are timed out, kicked, banned, or unbanned, create/append to
an entry in Nuclino and a thread in Discord.

Posting to forum channels required an upgrade to discord.py 2.0, which
also means that Python 3.8 is now the minimum supported version (the
code itself currently needs Python 3.9). Since discord.py now has
built-in support for commands, the commands module has been migrated to
use it. However, it's currently unclear if it actually works (I get
errors complaining about missing commands).

Also introduces a new dependency-injection-style system for managing
the bot's modules, and migrates the 2 existing modules that are
currently in use.